refactor(plugins): move the config-defaults seeder out of the loader
seedConfigDefaults is a pure function — sixteen lines, no `this`, no injected
dependency — that lived inside a 1713-line service module with 31 imports.
Reaching it meant loading the whole loader graph: BullMQ, ModuleRef, the sandbox
worker host, plugin storage, the hook manager. Its own spec block paid that
price to test schema defaulting.

It now sits in config-defaults.util.ts with its tests colocated beside it,
matching how search-provider-registration and webhook-subscribe were already
lifted out of the same file.

Behaviour is unchanged: the function moved verbatim, the four tests moved
verbatim, and the suite total is identical (4011) because nothing was added —
the loader spec drops from 56 to 52 and the new file accounts for the rest. The
loader's PluginConfigSchema import goes with it, since seeding was its only use.

Nothing imported the symbol through the core/plugins barrel, so no re-export is
added; narrowing that surface is part of the point.
